refactor(fuse): extract replay.rs and slim lib.rs to ~74 LoC crate root
- Create crates/fuse/src/replay.rs: replay_for_vault (pub), resolve_folder_key,
  resolve_folder_key_cached, fetch_merge_publish_parent, publish_child_folder_metadata,
  replay_mkdir_entry, replay_upload_entry, decrypt_journal_name (pub(crate)),
  plus all replay tests (T-45-06, T-45-07, REQ-4, REQ-5, T-45-05, F2)
- Slim lib.rs from ~2128 to ~74 LoC production declarations + re-exports;
  handler_harness_tests and durability_characterization_tests stay per RESEARCH Pitfall 3
- Fix crate::NETWORK_TIMEOUT → crate::runtime::NETWORK_TIMEOUT in read_ops.rs
- Update durability_characterization_tests to use crate::replay::decrypt_journal_name
  (Rule 3 fix: decrypt_journal_name moved to replay.rs as pub(crate))
